Add logic for resource overrides
This is the companion commit to SensibilityTestbed@7539649aa35f9fb92e263f0832032d71ad0264b8.
It adds an `overrides` parameter to the `benchmark_resources/main`
function so that the installer can pass along resource value overrides
that the user specified on the command line. Machine resources are
benchmarked as usual, but if the user specified values for resources,
then the benchmarked values are discarded and replaced by the user input.

In order to keep things in sync, a notification is output if the
benchmark script benchmarks a resource for which there is no override
flag on the command line.